Learning frontend and ML technologies.
TruthGuard is a powerful tool designed to combat fake news, offering a user-friendly interface with real-time alerts.
Backed by transparent assessments and a community-driven reporting feature, it empowers users to confidently
navigate the digital news landscape. In a world filled with misinformation, TruthGuard stands as a reliable defense for truth-seekers.